packages/workers/provisioner/pyproject.toml
[tool.poetry]
name = "insighthub-worker-provisioner"
version = "0.1.0"
description = "Workspace provisioner worker - creates RAG infrastructure for workspaces"
authors = ["Nate Stott <nate.stott@pm.me>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
qdrant-client = "^1.7.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
insighthub-shared = {path = "../../shared/python", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.0.0"
mypy = "^1.8.0"
ruff = "^0.2.0"
isort = "^5.13.0"
testcontainers = "^4.0.0"
pytest-asyncio = "^0.23.0"
types-psycopg2 = "^2.9.0"

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "pika.*",
    "psycopg2.*",
    "qdrant_client.*",
    "testcontainers.*",
    "shared.*",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"